requirement.txt
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
openai>=1.0.0
python-dotenv>=0.21.0
requests>=2.28.0
GitPython>=3.1.30
pyyaml